requirements.txt
email-validator>=2.2.0
flask==2.3.3
flask-sqlalchemy>=3.1.1
flask-login==0.6.3
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
requests-oauthlib>=1.3.1
google-generativeai==0.5.4
gunicorn>=23.0.0
requests==2.31.0
sqlalchemy==2.0.23
python-dotenv>=1.0.0 